requirements.windows.txt
adi-reader
pandas
numpy